refactor(llama): move TokenUtf8Buffer to util
