+++ llama-rs/src/lib.rs
@@ -22,6 +22,7 @@ pub use ggml::Type as ElementType;
 pub use inference_session::{
     InferenceSession, InferenceSessionParameters, InferenceSnapshot, ModelKVMemoryType,
 };
+pub use util::TokenUtf8Buffer;
 
 /// The end of text token.
 pub const EOT_TOKEN_ID: TokenId = 2; // Hardcoded (for now?)
@@ -1349,84 +1350,5 @@ impl Vocabulary {
         res.reverse();
 
         Ok(res)
-    }
-}
-
-/// Used to buffer incoming tokens until they produce a valid string of UTF-8 text.
-///
-/// Tokens are *not* valid UTF-8 by themselves. However, the LLM will produce valid UTF-8
-/// from multiple tokens. This helps alleviate that issue.
-#[derive(Clone, PartialEq, Default)]
-pub struct TokenUtf8Buffer(Vec<u8>);
-impl TokenUtf8Buffer {
-    /// Create a new buffer.
-    pub const fn new() -> Self {
-        Self(vec![])
-    }
-
-    /// Add a token to the buffer. If the buffer contains a valid string of UTF-8 text,
-    /// it is returned and the buffer is cleared for next use.
-    pub fn push(&mut self, token: &[u8]) -> Option<String> {
-        self.0.extend_from_slice(token);
-        match std::str::from_utf8(&self.0) {
-            Ok(s) => {
-                let out = s.to_owned();
-                self.0 = vec![];
-                Some(out)
-            }
-            Err(..) => {
-                for i in 1..self.0.len() {
-                    let slice = &self.0[i..];
-                    if slice.is_empty() {
-                        break;
-                    }
-
-                    if let Ok(s) = std::str::from_utf8(slice) {
-                        let out = s.to_owned();
-                        self.0 = vec![];
-                        return Some(out);
-                    }
-                }
-                None
-            }
-        }
-    }
-
-    /// Adapt a `&str` callback so that it can be used in a `&[u8]` context.
-    fn adapt_callback<'a, E: std::error::Error + 'static>(
-        mut callback: impl FnMut(&str) -> Result<(), E> + 'a,
-    ) -> impl FnMut(&[u8]) -> Result<(), E> + 'a {
-        let mut buffer = Self::new();
-        move |token| match buffer.push(token) {
-            Some(tokens) => callback(&tokens),
-            None => Ok(()),
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_valid_utf8() {
-        let mut buffer = TokenUtf8Buffer::new();
-        assert_eq!(buffer.push(b"hello").as_deref(), Some("hello"));
-        assert_eq!(buffer.push(&[0xE2, 0x82, 0xAC]).as_deref(), Some("€"));
-    }
-
-    #[test]
-    fn test_partial_utf8() {
-        let mut buffer = TokenUtf8Buffer::new();
-        assert_eq!(buffer.push(&[0xE2, 0x82]).as_deref(), None);
-        assert_eq!(buffer.push(&[0xAC]).as_deref(), Some("€"));
-    }
-
-    #[test]
-    fn test_invalid_prelude_for_valid_utf8() {
-        let mut buffer = TokenUtf8Buffer::new();
-        assert_eq!(buffer.push(&[0xD8]).as_deref(), None);
-        assert_eq!(buffer.push(&[0xE2, 0x82]).as_deref(), None);
-        assert_eq!(buffer.push(&[0xAC]).as_deref(), Some("€"));
     }
 }
